requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.7
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.0.3
